Chore: update events and projects content section
