+++ app/events/events.ts
@@ -1,0 +1,229 @@
+export const pastEvents = [
+    {
+      title:
+        "TechCare: Project Dagtaan",
+      date: "15 Mar, 22 Mar, and 29 Mar 2025",
+      type: "Outreach Program",
+      location: "Kalipay Foundation",
+      image: "/images/dagtaan.jpg",
+      link: "https://project-dagtaan.vercel.app/",
+    },
+    {
+      title:
+        "Sonder 2025: Vietnam",
+      date: "22 Feb 2025",
+      type: "Conference",
+      location: "Virtual, GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/sonder-vn.jpg",
+      link: "https://www.facebook.com/share/p/1B4txGazfY/",
+    },
+    {
+      title:
+        "Sonder 2025: Indonesia",
+      date: "11 Jan 2025",
+      type: "Conference",
+      location: "Virtual, GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/sonder-ind.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-sonder-google-dev-world-tour-in-indonesia/cohost-gdg-on-campus-university-of-st-la-salle-bacolod-philippines"
+    },
+    {
+      title:
+        "SONDER: International GDG on Campus Community Building - Study Group 2024",
+      date: "20 Dec 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/sonder.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-sonder-international-gdg-on-campus-community-building-study-group-2024/cohost-gdg-on-campus-university-of-st-la-salle-bacolod-philippines",
+    },
+    {
+      title: "View-Source Networking Session",
+      date: "14 Dec 2024",
+      type: "Conference",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/viewsource.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-view-source-networking-session/",
+    },
+    {
+      title: "Core Team - Team Building Program",
+      date: "9 Nov 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/coreteam.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-core-team-team-building-program/",
+    },
+    {
+      title: "1st General Assembly & Chapter Characters Info Session",
+      date: "8 Nov 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/assembly.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-1st-general-assembly-amp-chapter-characters-info-session/",
+    },
+    {
+      title: "Public Relations Department Study Group & Working Session",
+      date: "30 Oct 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/pr.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-public-relations-department-study-group-amp-working-session/",
+    },
+    {
+      title: "Technology Department Study Group & Working Session",
+      date: "30 Oct 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/tech.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-technology-department-study-group-amp-working-session/",
+    },
+    {
+      title: "Events Department Study Group & Working Session",
+      date: "30 Oct 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/events.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-events-department-study-group-amp-working-session/",
+    },
+    {
+      title: "Membership Transitioning & General Assembly",
+      date: "25 Oct 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/general.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-membership-transitioning-amp-general-assembly/",
+    },
+    {
+      title: "Core Team Working Session & Planning",
+      date: "17 Oct 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/coreteamass.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-core-team-working-session-amp-planning/",
+    },
+    {
+      title: "Hack La Salle 2024 Info Session & Ideathon Community Mixer",
+      date: "16 Oct 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/hack.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-hack-la-salle-2024-info-session-amp-ideathon-community-mixer/",
+    },
+    {
+      title: "Internal Strategic Meeting & Bevy Workshop",
+      date: "11 Oct 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/chill.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-internal-strategic-meeting-amp-bevy-workshop/",
+    },
+    {
+      title: "Hack La Salle 2024 - Kickoff Event & Creatives Workshop",
+      date: "24 Sept 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/hacklasalle.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-hack-la-salle-2024-kickoff-event-amp-creatives-workshop/",
+    },
+    {
+      title: "Game On! USLS 2024",
+      date: "20 Sept 2024",
+      type: "Hackathon",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/gameon.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-game-on-usls-2024/",
+    },
+    {
+      title: "Hack La Salle 2024",
+      date: "16 Sept 2024",
+      type: "Hackathon",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/hack2024.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-hack-la-salle-2024/",
+    },
+    {
+      title: "University Club Fair - The Capitol Booth & Series of Events",
+      date: "9 Sept 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/capitol.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-university-club-fair-the-capitol-booth-amp-series-of-events/",
+    },
+    {
+      title: "Community Events & Lead Workshop, Working Session",
+      date: "26 Aug 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/community.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-community-events-amp-lead-workshop-working-session/",
+    },
+    {
+      title: "Mastering UI/UX Etiquette Workshop",
+      date: "26 Aug 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/mastering.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-mastering-uiux-etiquette-workshop/",
+    },
+    {
+      title: "Character Design Workshop",
+      date: "24 Aug 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/chardesign.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-character-design-workshop/",
+    },
+    {
+      title: "Slack Workshop & Volunteers Onboarding Session",
+      date: "17 Aug 2024",
+      type: "Workshop / Study Group",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/slack.png",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-slack-workshop-amp-volunteers-onboarding-session/",
+    },
+    {
+      title: "College Freshmen Orientation: Club Booth & Info Session",
+      date: "3 Aug 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/breakout.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-college-freshmen-orientation-club-booth-amp-info-session/",
+    },
+    {
+      title:
+        "University HEU Orientation Program - Club Promotion & Info Session",
+      date: "1 Aug 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/freshmen.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-university-heu-orientation-program-club-promotion-amp-info-session/",
+    },
+    {
+      title: "Next: Navigating Extraordinary Talents",
+      date: "1 Aug 2024",
+      type: "Info Session",
+      location:
+        "GDG on Campus University of St. La Salle - Bacolod, Philippines",
+      image: "/images/info.jpg",
+      link: "https://gdg.community.dev/events/details/google-gdg-on-campus-university-of-st-la-salle-bacolod-philippines-presents-next-navigating-extraordinary-talents/",
+    },
+];
